* Change popup dialog to not use textarea nor include default 'failure-details' message. * displayError() now populates dialog HTML inside <p> tags * added popup-ok message
The only places I see displayError() currently used is in handling failure cases
for image uploads, and the two areas I've implemented for login failures. I'm not sure
what impact these changes will have on image upload failures - I suspect the
'failure-details' message was intended to help wiht the upload failures. If that's
the case, perhaps it can be prepended to the error message prior to display only in
that use case, rather than ubiquitously as it seemed gratuitous for login failure
cases, and I suspect it will be overkill for many other uses of displayError() in
the future.