Stop escaping pipes in admintree names
I updated the API to handle this for you - which makes the API
more friendly to current API users and also makes the escaping
headache in the app a little less severe than it already is.
The API now automagically knows the difference between a | used
for separating adm levels and a | inside of wikitext.
As such, this has a dependency on r862 getting deployed to the
API servers.