throw away countries which are not described in campaigns-data.js
when campaigns get added we automatically pick them up and list them
on the first page. However if the campaigns-data.js has not been updated
when a photo is uploaded an error is thrown as there is no way to know
what license and template to upload them under.

instead throw these campaigns away so they are not surfaced in the app
log to console so this problem can be identified. this signals that
you need to run python scripts/generate_campaigns.py