fix search to be more useful
currently if you have given your location a bounding box is set and
all searches from the list view are now based on that bounding box

it's actually more easy to ignore this. If campaign is set it will
search within that campaign. If not it will search across the whole
api for monuments. I think this is much more useful and aids discovery