Mark camera and location features as optional in the manifest.
Should help with devices listing the app as incompatible in the Google Play store.

Camera is optional because we can pull from your gallery, which might be synced with other sources.
Also "camera" means "front-facing camera". :P
Location services are optional as well, and the app should function without them.